+++ src/health/health.controller.ts
@@ -1,26 +1,20 @@
 import { Controller, Get } from '@nestjs/common';
-import {
-  HealthCheck,
-  HealthCheckService,
-  HttpHealthIndicator,
-} from '@nestjs/terminus';
-import { ConfigService } from '@nestjs/config';
-import Config from '../config/config.d';
+import { HealthCheck } from '@nestjs/terminus';
 
 @Controller('health')
 export class HealthController {
-  constructor(
-    private config: ConfigService,
-    private health: HealthCheckService,
-    private dns: HttpHealthIndicator,
-  ) {}
-
   @Get()
   @HealthCheck()
   check() {
-    const baseURL = this.config.get<Config>('confluence.baseURL');
-    return this.health.check([
-      async () => this.dns.pingCheck('Atlassian API', `${baseURL}`),
-    ]);
+    // return this.health.check([
+    //   async () => this.dns.pingCheck('Atlassian API', `${baseURL}`),
+    // ]);
+
+    return {
+      status: 'ok',
+      info: { Konviw: { status: 'up' } },
+      error: {},
+      details: { Konviw: { status: 'up' } },
+    };
   }
 }